Add logo, add left and right rotation
