+++ www/src/js/views/tetris/TetrisGame.scss
@@ -36,7 +36,7 @@
 }
 
 .overlayContent {
-  width: 20rem;
+  width: 40rem;
   margin: 6rem auto 0;
   text-align: center;
 }
@@ -45,10 +45,12 @@
   margin-left: 1rem;
 
   section {
-    margin-bottom: 1.5rem;
+    margin-bottom: 2rem;
+    line-height: 1;
   }
 
   h3 {
+    opacity: 0.7;
     margin-bottom: 0;
   }
 }
