+++ www/src/js/views/tetris/TetrisLogo.scss
@@ -1,0 +1,10 @@
+.wrapper {
+  display: flex;
+  justify-content: center;
+}
+
+.logo {
+  display: inline-block;
+  font-weight: bold;
+  text-align: left;
+}
